Added AudioCompression options

+++ Compressor.cs
@@ -27,6 +27,7 @@ namespace CryCompressor
         readonly ConcurrentQueue<string> fileQueue = new ConcurrentQueue<string>();
         readonly ConcurrentQueue<string> videoQueue = new ConcurrentQueue<string>();
         readonly ConcurrentQueue<string> imageQueue = new ConcurrentQueue<string>();
+        readonly ConcurrentQueue<string> audioQueue = new ConcurrentQueue<string>();
 
         readonly ConcurrentQueue<string> errorQueue = new ConcurrentQueue<string>();
         readonly ConcurrentQueue<string> warnQueue = new ConcurrentQueue<string>();
@@ -35,6 +36,8 @@ namespace CryCompressor
         readonly Dictionary<int, bool> videoParams = new Dictionary<int, bool>();
         readonly SemaphoreSlim imageParamsSemaphore = new SemaphoreSlim(1);
         readonly Dictionary<int, bool> imageParams = new Dictionary<int, bool>();
+        readonly SemaphoreSlim audioParamsSemaphore = new SemaphoreSlim(1);
+        readonly Dictionary<int, bool> audioParams = new Dictionary<int, bool>();
 
         public Compressor(Configuration config, CancellationToken token)
         {
@@ -43,6 +46,7 @@ namespace CryCompressor
 
             for (int i = 0; i < config.VideoCompression.ParametersPriorityList.Length; i++) videoParams.Add(i, false);
             for (int i = 0; i < config.ImageCompression.ParametersPriorityList.Length; i++) imageParams.Add(i, false);
+            for (int i = 0; i < config.AudioCompression.ParametersPriorityList.Length; i++) audioParams.Add(i, false);
         }
 
         public Task Start()
@@ -61,6 +65,7 @@ namespace CryCompressor
             // Start up workers
             for (int i = 0; i < configuration.VideoCompression.MaxConcurrentWorkers; i++) _ = Task.Run(VideoWorker, tkn);
             for (int i = 0; i < configuration.ImageCompression.MaxConcurrentWorkers; i++) _ = Task.Run(ImageWorker, tkn);
+            for (int i = 0; i < configuration.AudioCompression.MaxConcurrentWorkers; i++) _ = Task.Run(AudioWorker, tkn);
             _ = Task.Run(FileWorker, tkn);
 
             // Start up progress logger
@@ -69,6 +74,7 @@ namespace CryCompressor
             // Start filtering files and queueing them for work
             var vidExtensions = configuration.VideoExtensions.Select(x => "." + x.ToLower().Trim().GetExtensionWithoutDot()).ToArray();
             var imgExtensions = configuration.ImageExtensions.Select(x => "." + x.ToLower().Trim().GetExtensionWithoutDot()).ToArray();
+            var audExtensions = configuration.AudioExtensions.Select(x => "." + x.ToLower().Trim().GetExtensionWithoutDot()).ToArray();
             foreach (var f in files)
             {
                 var extension = Path.GetExtension(f).ToLower().Trim();
@@ -83,6 +89,11 @@ namespace CryCompressor
                 {
                     // IMAGE
                     imageQueue.Enqueue(f);
+                }
+                else if (configuration.AudioCompression.CompressAudio && audExtensions.Contains(extension) && size >= configuration.AudioCompression.MinSize)
+                {
+                    // AUDIO
+                    audioQueue.Enqueue(f);
                 }
                 else
                 {
@@ -130,7 +141,7 @@ namespace CryCompressor
         }
 
         readonly static Random rng = new();
-        string GetPath(string filename, string extension = null)
+        (string destination_unchanged, string destination_modified) GetPath(string filename, string extension = null)
         {
             extension = extension?.GetExtensionWithoutDot();
 
@@ -150,9 +161,10 @@ namespace CryCompressor
                 if (fileExtension.ToLower() != extension.ToLower()) fileNameNoext += $"({rng.Next(0, 1000)}-{fileExtension})"; 
             }
 
-            var destination = Path.Combine(destinationDirectory, fileNameNoext + "." + extension);
+            var destination_unchanged = Path.Combine(destinationDirectory, Path.GetFileName(filename));
+            var destination_modified = Path.Combine(destinationDirectory, fileNameNoext + "." + extension);
 
-            return destination;
+            return (destination_unchanged, destination_modified);
         }
 
         async Task VideoWorker()
@@ -167,8 +179,9 @@ namespace CryCompressor
 
                     // choose parameters based on priority list
                     var (pIndex, parameters) = await TakeVideoParameters();
-                    var dst = GetPath(f, parameters.Extension);
+                    var (dst_orig, dst) = GetPath(f, parameters.Extension);
 
+                    string output = "";
                     try
                     {
                         // convert
@@ -185,7 +198,6 @@ namespace CryCompressor
 
                         reader.Dispose();
 
-                        string output = "";
                         p = FFmpegWrapper.ExecuteCommand("ffmpeg", $"-hide_banner -i \"{f}\" {parameters.Parameters} \"{dst}\" -y");
                         p.ErrorDataReceived += (s, data) =>
                         {
@@ -213,10 +225,10 @@ namespace CryCompressor
                     }
                     catch (Exception ex)
                     {
-                        // in rare case it hasn't been copied over, copy it
-                        if (!File.Exists(dst)) File.Copy(f, dst, true);
+                        // in rare case it hasn't been copied over, copy it (use original filename for destination now instead)
+                        if (!File.Exists(dst)) File.Copy(f, dst_orig, true);
 
-                        errorQueue.Enqueue($"Failed to convert video '{f}': {ex.Message}");
+                        errorQueue.Enqueue($"Failed to convert video '{f}': {ex.Message}" + (output.Length > 0 ? $"\n\nFFmpeg output:\n{output}" : ""));
                     }
                     finally
                     {
@@ -244,8 +256,9 @@ namespace CryCompressor
                     // choose parameters based on priority list
                     var (pIndex, parameters) = await TakeImageParameters();
 
-                    var dst = GetPath(f, parameters.Extension);
+                    var (dst_orig, dst) = GetPath(f, parameters.Extension);
 
+                    string output = "";
                     try
                     {
                         // convert
@@ -255,7 +268,6 @@ namespace CryCompressor
 
                         reader.Dispose();
 
-                        string output = "";
                         p = FFmpegWrapper.ExecuteCommand("ffmpeg", $"-hide_banner -i \"{f}\" {parameters.Parameters} \"{dst}\" -y");
                         p.ErrorDataReceived += (s, data) =>
                         {
@@ -283,14 +295,84 @@ namespace CryCompressor
                     }
                     catch (Exception ex)
                     {
-                        // in rare case it hasn't been copied over, copy it
-                        if (!File.Exists(dst)) File.Copy(f, dst, true);
+                        // in rare case it hasn't been copied over, copy it (use original filename for destination now instead)
+                        if (!File.Exists(dst)) File.Copy(f, dst_orig, true);
 
-                        errorQueue.Enqueue($"Failed to convert image '{f}': {ex.Message}");
+                        errorQueue.Enqueue($"Failed to convert image '{f}': {ex.Message}" + (output.Length > 0 ? $"\n\nFFmpeg output:\n{output}" : ""));
                     }
                     finally
                     {
                         await ReleaseImageParameters(pIndex);
+
+                        if (p != null && !p.HasExited) p.Kill();
+
+                        Interlocked.Increment(ref filesProcessed);
+                    }
+                }
+                else if (queuedone) break;
+
+                await Task.Delay(10);
+            }
+        }
+
+        async Task AudioWorker()
+        {
+            while (!tkn.IsCancellationRequested)
+            {
+                if (audioQueue.TryDequeue(out string f))
+                {
+                    Process p = null;
+
+                    // choose parameters based on priority list
+                    var (pIndex, parameters) = await TakeAudioParameters();
+
+                    var (dst_orig, dst) = GetPath(f, parameters.Extension);
+
+                    string output = "";
+                    try
+                    {
+                        // convert
+                        using var reader = new VideoReader(f);
+
+                        await reader.LoadMetadataAsync();
+
+                        reader.Dispose();
+                        
+                        p = FFmpegWrapper.ExecuteCommand("ffmpeg", $"-hide_banner -i \"{f}\" {parameters.Parameters} \"{dst}\" -y");
+                        p.ErrorDataReceived += (s, data) =>
+                        {
+                            if (data.Data == null) return;
+                            output += data.Data + "\n";
+                        };
+
+                        await p.WaitForExitAsync(tkn);
+                        var code = p.ExitCode;
+
+                        // validate result
+                        var result = new FileInfo(dst);
+                        if (result.Length < 1000 || code != 0)
+                        {
+                            File.Delete(dst);
+                            File.Copy(f, dst, true);
+                            throw new Exception($"Audio conversion failed. (Code: {code}). Output:\n{output}");
+                        }
+                        if (configuration.DeleteResultIfBigger && result.Length > new FileInfo(f).Length)
+                        {
+                            File.Delete(dst);
+                            File.Copy(f, dst, true);
+                            warnQueue.Enqueue($"Converted audio was larger than original, overwriting it. ('{f}')");
+                        }
+                    }
+                    catch (Exception ex)
+                    {
+                        // in rare case it hasn't been copied over, copy it (use original filename for destination now instead)
+                        if (!File.Exists(dst)) File.Copy(f, dst_orig, true);
+
+                        errorQueue.Enqueue($"Failed to convert audio '{f}': {ex.Message}" + (output.Length > 0 ? $"\n\nFFmpeg output:\n{output}" : ""));
+                    }
+                    finally
+                    {
+                        await ReleaseAudioParameters(pIndex);
 
                         if (p != null && !p.HasExited) p.Kill();
 
@@ -311,7 +393,7 @@ namespace CryCompressor
                 {
                     try
                     {
-                        var dst = GetPath(f);
+                        var (_, dst) = GetPath(f);
                         File.Copy(f, dst, true);
                     }
                     catch (Exception ex)
@@ -411,6 +493,48 @@ namespace CryCompressor
             finally
             {
                 imageParamsSemaphore.Release();
+            }
+        }
+
+        async Task<(int Index, ParametersObject Parameters)> TakeAudioParameters()
+        {
+            // pick the first parameters available - if last, take last one.
+
+            await audioParamsSemaphore.WaitAsync();
+
+            try
+            {
+                for (int i = 0; i < configuration.AudioCompression.ParametersPriorityList.Length; i++)
+                {
+                    if (!audioParams[i])
+                    {
+                        audioParams[i] = true;
+                        return (i, configuration.AudioCompression.ParametersPriorityList[i]);
+                    }
+                }
+
+                // always return last one if all else is taken
+                return (
+                    configuration.AudioCompression.ParametersPriorityList.Length - 1,
+                    configuration.AudioCompression.ParametersPriorityList.Last());
+            }
+            finally
+            {
+                audioParamsSemaphore.Release();
+            }
+        }
+
+        async Task ReleaseAudioParameters(int index)
+        {
+            await audioParamsSemaphore.WaitAsync();
+
+            try
+            {
+                audioParams[index] = false;
+            }
+            finally
+            {
+                audioParamsSemaphore.Release();
             }
         }
         #endregion
